imgsys: fix incompatible data size between 32bit/64bit
In ioctl command MTKDIP_IOC_ADD_IOVA/MTKDIP_IOC_DEL_IOVA, the
sizeof(struct fd_tbl) is not the same between 32bit/64bit due to
fd_tbl.fds is a pointer, in 32bit environment, the pointer length is 8,
but in 64bit environment, the pointer length is 4. It cause the ioctl
command MTKDIP_IOC_ADD_IOVA/MTKDIP_IOC_DEL_IOVA cannot be found.

To fix this, we use fix length 'uint64_t' to store a pointer address
instead of unsigned int*. We also remove struct fd_tbl_compat which is
no longer be used.

Change-Id: Ie71b68768ebf129654994df87c4e3d62696432c2
Signed-off-by: Efang Chung <efang.chung@mediatek.com>