mtk-camera: Add g_ctrl V4L2_CID_MIN_BUFFERS_FOR_CAPTURE
This patch adds V4L2_CID_MIN_BUFFERS_FOR_CAPTURE for querying
min buffers. This is a reference for the app to determine how
many buffers should be allocated for a pipeline.

The value of min_buffers_needed for YUV sensor is set to 3,
which is the minimum number of buffers needed before calling
start_streaming. The reference number of buffers for app is
set to min_buffers_needed + 1.
+----------+    +---------+    +---------+
| Hardware | -> | Driver  | -> |  User   |
| (2 Bufs) | <- | (1 Buf) | <- | (1 Buf) |
+----------+    +---------+    +---------+

In addition, the value of min_buffer_needed is determined after
the driver requests to the daemon for the information, so the
mtk_camera_init_stream function should be called before querying
controls.

Change-Id: If3c4a3a5dbc797058a94fefbf1961288385ef9a3
Signed-off-by: Andy.Hsieh <andy.hsieh@mediatek.com>