mtk-camera: Make driver functions static
This patch modifies the mtk-camera driver in order to improve
code organization and limit the visibility of certain functions.
By making these functions static, their scope is restricted to
the source file where they are defined. This helps to enhance
code encapsulation and minimize potential naming conflicts
with other modules.

Change-Id: I0387dafb3ee473da8709cc862b78907444e1b50c
Signed-off-by: Andy.Hsieh <andy.hsieh@mediatek.com>